Partial revert of 109260442c483d6181d25f1a2c17511f6c23193e
It turns for disassembly we don't really need to keep track of flags set
by Choice*. Branch operation is really binary, so for ChoiceIdx we
have a few Branch operations to test all the variants. It's still
unknown how the first vararg of Branch is chosen, but it's not important
for disassembly.